crawl-data-web/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
lxml==4.9.3
pandas==2.1.3
sqlalchemy==2.0.23
python-dotenv==1.0.0
scrapy==2.11.0
fake-useragent==1.4.0
aiohttp==3.9.1
asyncio==3.4.3
urllib3==2.1.0
